Create csc test package(s)
