+++ src/test/java/redis/clients/jedis/csc/JedisClusterClientSideCacheTest.java
@@ -1,38 +1,33 @@
-package redis.clients.jedis;
+package redis.clients.jedis.csc;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Set;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
+
 import org.apache.commons.pool2.impl.GenericObjectPoolConfig;
 import org.hamcrest.Matchers;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import redis.clients.jedis.csc.DefaultClientSideCacheConfig;
-import redis.clients.jedis.util.MapCSC;
+import redis.clients.jedis.Connection;
+import redis.clients.jedis.ConnectionPoolConfig;
+import redis.clients.jedis.DefaultJedisClientConfig;
+import redis.clients.jedis.HostAndPort;
+import redis.clients.jedis.JedisClientConfig;
+import redis.clients.jedis.JedisCluster;
+import redis.clients.jedis.JedisClusterTestBase;
+import redis.clients.jedis.csc.util.MapCSC;
 
-public class JedisPooledClientSideCacheTest {
+public class JedisClusterClientSideCacheTest extends JedisClusterTestBase {
 
-  protected static final HostAndPort hnp = HostAndPorts.getRedisServers().get(1);
-
-  protected Jedis control;
-
-  @Before
-  public void setUp() throws Exception {
-    control = new Jedis(hnp, DefaultJedisClientConfig.builder().password("foobared").build());
-    control.flushAll();
-  }
-
-  @After
-  public void tearDown() throws Exception {
-    control.close();
-  }
+  private static final Set<HostAndPort> hnp = Arrays.asList(nodeInfo1, nodeInfo2, nodeInfo3).stream().collect(Collectors.toSet());
 
   private static final Supplier<JedisClientConfig> clientConfig
-      = () -> DefaultJedisClientConfig.builder().resp3().password("foobared").build();
+      = () -> DefaultJedisClientConfig.builder().resp3().password("cluster").build();
 
   private static final Supplier<GenericObjectPoolConfig<Connection>> singleConnectionPoolConfig
       = () -> {
@@ -43,10 +38,10 @@ public class JedisPooledClientSideCacheTest {
 
   @Test
   public void simple() {
-    try (JedisPooled jedis = new JedisPooled(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC()))) {
-      control.set("foo", "bar");
+    try (JedisCluster jedis = new JedisCluster(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC()))) {
+      jedis.set("foo", "bar");
       assertEquals("bar", jedis.get("foo"));
-      control.del("foo");
+      jedis.del("foo");
       assertThat(jedis.get("foo"), Matchers.oneOf("bar", null)); // ?
     }
   }
@@ -54,12 +49,13 @@ public class JedisPooledClientSideCacheTest {
   @Test
   public void simpleWithSimpleMap() {
     HashMap<Long, Object> map = new HashMap<>();
-    try (JedisPooled jedis = new JedisPooled(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC(map)), singleConnectionPoolConfig.get())) {
-      control.set("foo", "bar");
+    try (JedisCluster jedis = new JedisCluster(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC(map)),
+        singleConnectionPoolConfig.get())) {
+      jedis.set("foo", "bar");
       assertThat(map, Matchers.aMapWithSize(0));
       assertEquals("bar", jedis.get("foo"));
       assertThat(map, Matchers.aMapWithSize(1));
-      control.del("foo");
+      jedis.del("foo");
       assertThat(map, Matchers.aMapWithSize(1));
       assertEquals("bar", jedis.get("foo"));
       assertThat(map, Matchers.aMapWithSize(1));
@@ -72,10 +68,10 @@ public class JedisPooledClientSideCacheTest {
 
   @Test
   public void flushAll() {
-    try (JedisPooled jedis = new JedisPooled(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC()))) {
-      control.set("foo", "bar");
+    try (JedisCluster jedis = new JedisCluster(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC()))) {
+      jedis.set("foo", "bar");
       assertEquals("bar", jedis.get("foo"));
-      control.flushAll();
+      jedis.flushAll();
       assertThat(jedis.get("foo"), Matchers.oneOf("bar", null)); // ?
     }
   }
@@ -83,12 +79,13 @@ public class JedisPooledClientSideCacheTest {
   @Test
   public void flushAllWithSimpleMap() {
     HashMap<Long, Object> map = new HashMap<>();
-    try (JedisPooled jedis = new JedisPooled(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC(map)), singleConnectionPoolConfig.get())) {
-      control.set("foo", "bar");
+    try (JedisCluster jedis = new JedisCluster(hnp, clientConfig.get(), new DefaultClientSideCacheConfig(new MapCSC(map)),
+        singleConnectionPoolConfig.get())) {
+      jedis.set("foo", "bar");
       assertThat(map, Matchers.aMapWithSize(0));
       assertEquals("bar", jedis.get("foo"));
       assertThat(map, Matchers.aMapWithSize(1));
-      control.flushAll();
+      jedis.flushAll();
       assertThat(map, Matchers.aMapWithSize(1));
       assertEquals("bar", jedis.get("foo"));
       assertThat(map, Matchers.aMapWithSize(1));
